Add textbuffers to editor tests for verification purposes.
